[Charlene] separate interpreter from the controller; add code for exporting files from diff modules
